requirements.txt
black==22.8.0
click==8.1.3
dnspython==2.2.1
Flask==2.2.2
Flask-PyMongo==2.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mongoengine==0.24.2
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
pymongo==4.2.0
python-dotenv==0.21.0
tomli==2.0.1
Werkzeug==2.2.2
